invoice

+++ apps/sendingRequest/views.py
@@ -11,6 +11,7 @@ from rest_framework.views import APIView
 from apps.users.models import Member
 from .models import SendingRequest
 from .serializers import SendingRequestSerializer, AdminSendingRequestSerializer
+from ..invoice.views import invoice_sending_request_post
 
 # Views for Sending Request
 
@@ -38,7 +39,8 @@ body_parameters = openapi.Schema(
         # Détails du colis (Cargo Details)
         'cargo_type': openapi.Schema(
             type=openapi.TYPE_STRING,
-            description="Type of cargo being sent"
+            description="Type of cargo being sent",
+            enum=["container", "pallets_boxes", "bulk_fret", "vehicle", "animals", "furniture_tools", "other"]
         ),
         'weight': openapi.Schema(
             type=openapi.TYPE_NUMBER,
@@ -104,21 +106,21 @@ body_parameters = openapi.Schema(
         'priority': openapi.Schema(
             type=openapi.TYPE_STRING,
             default='medium',
-            description="Priority level of the request"
+            description="Priority level of the request",
+            enum=["medium", "high"],
         ),
 
         # Mode de paiement (Payment Method)
         'payment_method': openapi.Schema(
             type=openapi.TYPE_STRING,
             description="Payment method for the request",
-            nullable=True
         ),
     },
     required=[
         'client', 'recipient_name', 'recipient_email', 'recipient_phone',
         'cargo_type', 'weight', 'dimensions', 'quantity',
         'pickup_location', 'pickup_date_time',
-        'delivery_location', 'delivery_date_time'
+        'delivery_location', 'delivery_date_time', 'payment_method'
     ]
 )
 
@@ -138,13 +140,13 @@ class SendingRequestView(APIView):
         tags=[tags]
     )
     def post(self, request):
+        user = request.user
         try:
-            client = Member.objects.get(user_ptr=request.user)
+            client = Member.objects.get(user_ptr=user)
 
         except Member.DoesNotExist:
             return Response({"error": "You must be a client to perform this request"}, status=status.HTTP_403_FORBIDDEN)
 
-        user = request.user
         if user.role == "client" or user.role == "company":
             data = request.data.copy()  # Copier les données pour les modifier
             data['client'] = client.id  # Ajouter l'ID du driver
@@ -152,6 +154,7 @@ class SendingRequestView(APIView):
             serializer = SendingRequestSerializer(data=data)
             if serializer.is_valid():
                 serializer.save()
+                invoice_sending_request_post(serializer.data)
                 return Response(serializer.data, status=status.HTTP_201_CREATED)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
@@ -225,8 +228,39 @@ class SendingRequestDetailsView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @swagger_auto_schema(
+        operation_description="Relancer une demande annuler precedement (possibilite de mettre a jour les specifications)",
+        request_body=body_parameters,
+        responses={
+            200: openapi.Response("Request relaunched"),
+            403: openapi.Response("User unauthorized"),
+            404: openapi.Response("Request not found"),
+            500: openapi.Response("Internal Server Error"),
+        },
+        tags=[tags]
+
+    )
+    def post(self, request, pk):
+        # Logique de payment ou l'user doit donner 20% du prix de depart pour relancer un demande qu'il a annuler
+
+        sending_request = self.get_object(pk)
+        if request.data:
+            data = request.data.copy()
+            data['status'] = 'pending'
+            serializer = SendingRequestSerializer(sending_request, data=data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            sending_request.status = "pending"
+            sending_request.save()
+            serializer = SendingRequestSerializer(sending_request)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @swagger_auto_schema(
         operation_description="Delete a specific sending request",
         responses={
+            200: openapi.Response("Request cancelled"),
             204: openapi.Response("Request deleted successfully"),
             403: openapi.Response("User unauthorized"),
             404: openapi.Response("Request not found"),
@@ -236,9 +270,23 @@ class SendingRequestDetailsView(APIView):
     )
     def delete(self, request, pk):
         sending_request = self.get_object(pk)
-        if sending_request:
+        # Pour supprimer un demande deja annule
+        if sending_request and sending_request.status in ["cancelled", "rejected"]:
             sending_request.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
+        elif sending_request and sending_request.status == "in_progress":
+            sending_request.status = "cancelled"
+            sending_request.save()
+            # Utiliser le sérialiseur pour formater les données
+            serializer = SendingRequestSerializer(sending_request)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        elif sending_request.status in ["accepted", "in_progress"]:
+            # logique pour la deduction de 20% de l'argent deja envoye et l'attente pour recevoire le retour de l'argent
+            sending_request.status = "cancelled"
+            sending_request.save()
+            serializer = SendingRequestSerializer(sending_request)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+
         return Response({"error": f"Error for deleting request"},
                         status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
@@ -260,7 +308,7 @@ class AdminSendingRequestUpdateView(APIView):
             properties={
                 'status': openapi.Schema(
                     type=openapi.TYPE_STRING,
-                    enum=['pending', 'in_progress', 'completed', 'cancelled', 'rejected'],
+                    enum=['accepted', 'rejected'],
                     description="New status for the sending request")
             }),
         responses={
@@ -283,7 +331,16 @@ class AdminSendingRequestDetailsView(APIView):
     parser_classes = [JSONParser]
 
     @swagger_auto_schema(
-        operation_description="Afficher la liste des demandes",
+        operation_description="Afficher la liste des demandes avec filtres optionnels",
+        manual_parameters=[
+            openapi.Parameter(
+                'status',
+                openapi.IN_QUERY,
+                description="Filtrer les demandes par statut (ex: accepted, canceled)",
+                type=openapi.TYPE_STRING,
+                required=False
+            )
+        ],
         responses={
             200: openapi.Response("List of sending request", SendingRequestSerializer),
             403: openapi.Response("User unauthorized"),
@@ -291,7 +348,22 @@ class AdminSendingRequestDetailsView(APIView):
         tags=[tags]
     )
     def get(self, request):
-        send_requests = SendingRequest.objects.all()
+        valid_statuses = ['pending', 'accepted', 'canceled', 'completed', 'rejected', 'in_progress']
+        status_filter = request.query_params.get('status', None)
+
+        # Filtrer les demandes en fonction du statut
+        if status_filter:
+            # Si plusieurs statuts sont fournis (séparés par des virgules)
+            statuses = status_filter.split(',')
+            statuses = [stat for stat in statuses if stat in valid_statuses]
+            if not statuses:
+                return Response({"error": "Invalid status values"}, status=status.HTTP_400_BAD_REQUEST)
+            send_requests = SendingRequest.objects.filter(status__in=statuses)
+        else:
+            # Si aucun filtre n'est fourni, retourner toutes les demandes
+            send_requests = SendingRequest.objects.all()
+
+        # Sérialiser les résultats
         serializer = SendingRequestSerializer(send_requests, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -310,8 +382,8 @@ class ChiefFleetSendingRequestDetailsView(APIView):
     )
     def get(self, request):
         if request.user.role == "chief":
-            send_requests = SendingRequest.objects.filter(status="in_progress")
+            send_requests = SendingRequest.objects.filter(status="accepted")
             serializer = SendingRequestSerializer(send_requests, many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
         else:
-            return Response("User unauthorized", status=status.HTTP_403_FORBIDDEN)
+            return Response({"error": "User unauthorized"}, status=status.HTTP_403_FORBIDDEN)
